service_newPRAnnounce/add throwing Exc., time computing
